Removes check for getgroup and setgroup during compile
Heroku builds don't run as root and so they don't have permissions to
change the group a command runs as. As best as I can tell the three
commands that are installed (sv, runsv, and runsvdir) do not support
changing the command's group so this check is just getting in our way.